finished v1 of dialog
added a bunch of scripts to make a form of final fantasy style dialog system, there is one issue. in the dialogdisplay script there is a void update that needs an onclick or on keydown to advance the text. new eventsystem broke that and i have yet to figure it out.